requirements.txt
streamlit
tensorflow-cpu
mtcnn==0.1.0
tensorflow_cpu
keras==2.4.3
keras-vggface==0.6
keras_applications==1.0.8
